+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [0.17.12](https://github.com/IBM/platform-services-java-sdk/compare/0.17.11...0.17.12) (2021-02-25)
+
+
+### Bug Fixes
+
+* **IAM Policy Management:** re-gen service and add examples after recent API changes ([#74](https://github.com/IBM/platform-services-java-sdk/issues/74)) ([6ee9286](https://github.com/IBM/platform-services-java-sdk/commit/6ee9286069a9b2302d4c706e0d021b245ba8e4c4))
+
+## [0.17.11](https://github.com/IBM/platform-services-java-sdk/compare/0.17.10...0.17.11) (2021-02-25)
+
+
+### Bug Fixes
+
+* **User Management:** add examples ([#80](https://github.com/IBM/platform-services-java-sdk/issues/80)) ([2c152fb](https://github.com/IBM/platform-services-java-sdk/commit/2c152fb5f72a8151be7c8fc24d0a1bb325c9e87f))
+
 ## [0.17.10](https://github.com/IBM/platform-services-java-sdk/compare/0.17.9...0.17.10) (2021-02-18)